requirements.txt
fastapi
uvicorn
pydantic
python-dotenv

langchain
langchain-core
langchain-openai
langchain-groq
langchain-community
langgraph==0.2.14

streamlit
requests
tavily-python
typing-extensions